+++ pkg/controllers/report/utils/scanner.go
@@ -11,7 +11,6 @@ import (
 	enginecontext "github.com/kyverno/kyverno/pkg/engine/context"
 	"github.com/kyverno/kyverno/pkg/engine/context/resolvers"
 	"github.com/kyverno/kyverno/pkg/engine/response"
-	"github.com/kyverno/kyverno/pkg/event"
 	"github.com/kyverno/kyverno/pkg/registryclient"
 	"go.uber.org/multierr"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -24,7 +23,6 @@ type scanner struct {
 	informerCacheResolvers resolvers.ConfigmapResolver
 	excludeGroupRole       []string
 	config                 config.Configuration
-	eventGen               event.Interface
 }
 
 type ScanResult struct {
@@ -42,7 +40,6 @@ func NewScanner(
 	rclient registryclient.Client,
 	informerCacheResolvers resolvers.ConfigmapResolver,
 	config config.Configuration,
-	eventGen event.Interface,
 	excludeGroupRole ...string,
 ) Scanner {
 	return &scanner{
@@ -51,50 +48,8 @@ func NewScanner(
 		rclient:                rclient,
 		informerCacheResolvers: informerCacheResolvers,
 		config:                 config,
-		eventGen:               eventGen,
 		excludeGroupRole:       excludeGroupRole,
 	}
-}
-
-func generateSuccessEvents(log logr.Logger, ers ...*response.EngineResponse) (eventInfos []event.Info) {
-	for _, er := range ers {
-		logger := log.WithValues("policy", er.PolicyResponse.Policy, "kind", er.PolicyResponse.Resource.Kind, "namespace", er.PolicyResponse.Resource.Namespace, "name", er.PolicyResponse.Resource.Name)
-		if !er.IsFailed() {
-			logger.V(4).Info("generating event on policy for success rules")
-			e := event.NewPolicyAppliedEvent(event.PolicyController, er)
-			eventInfos = append(eventInfos, e)
-		}
-	}
-	return eventInfos
-}
-
-func generateFailEvents(log logr.Logger, ers ...*response.EngineResponse) (eventInfos []event.Info) {
-	for _, er := range ers {
-		eventInfos = append(eventInfos, generateFailEventsPerEr(log, er)...)
-	}
-	return eventInfos
-}
-
-func generateFailEventsPerEr(log logr.Logger, er *response.EngineResponse) []event.Info {
-	var eventInfos []event.Info
-	logger := log.WithValues(
-		"policy", er.PolicyResponse.Policy.Name,
-		"kind", er.PolicyResponse.Resource.Kind,
-		"namespace", er.PolicyResponse.Resource.Namespace,
-		"name", er.PolicyResponse.Resource.Name,
-	)
-	for i, rule := range er.PolicyResponse.Rules {
-		if rule.Status != response.RuleStatusPass && rule.Status != response.RuleStatusSkip {
-			eventResource := event.NewResourceViolationEvent(event.PolicyController, event.PolicyViolation, er, &er.PolicyResponse.Rules[i])
-			eventInfos = append(eventInfos, eventResource)
-			eventPolicy := event.NewPolicyFailEvent(event.PolicyController, event.PolicyViolation, er, &er.PolicyResponse.Rules[i], false)
-			eventInfos = append(eventInfos, eventPolicy)
-		}
-	}
-	if len(eventInfos) > 0 {
-		logger.V(4).Info("generating events for policy", "events", eventInfos)
-	}
-	return eventInfos
 }
 
 func (s *scanner) ScanResource(ctx context.Context, resource unstructured.Unstructured, nsLabels map[string]string, policies ...kyvernov1.PolicyInterface) map[kyvernov1.PolicyInterface]ScanResult {
@@ -120,14 +75,6 @@ func (s *scanner) ScanResource(ctx context.Context, resource unstructured.Unstru
 			}
 		}
 		results[policy] = ScanResult{response, multierr.Combine(errors...)}
-	}
-	for _, result := range results {
-		eventInfos := generateFailEvents(s.logger, result.EngineResponse)
-		s.eventGen.Add(eventInfos...)
-		if s.config.GetGenerateSuccessEvents() {
-			successEventInfos := generateSuccessEvents(s.logger, result.EngineResponse)
-			s.eventGen.Add(successEventInfos...)
-		}
 	}
 	return results
 }
